backend/pyproject.toml
[project]
name = "ai-web-agent-backend"
version = "0.1.0"
description = "Backend for MCP-based autonomous browser agent"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "playwright>=1.53.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "pydantic>=2.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
